requirements.txt
addict==2.4.0
annotated-types==0.7.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==3.0.4
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.58.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
kiss-icp==1.2.3
kiss-slam==0.0.2
kiwisolver==1.4.8
map_closures==2.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mypy_extensions==1.1.0
narwhals==1.41.1
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
numpy-quaternion==2024.0.8
open3d==0.19.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==14.0.0
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.0.6
zipp==3.23.0
